+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/w3c/DisabledW3cPerformanceTimeline.kt
@@ -1,0 +1,9 @@
+package com.atlassian.performance.tools.jiraactions.api.w3c
+
+/**
+ * Yields no entries.
+ * Has no overhead on CPU, memory or I/O.
+ */
+class DisabledW3cPerformanceTimeline : W3cPerformanceTimeline {
+    override fun record(): RecordedPerformanceEntries? = null
+}