+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/w3c/NavigationType.kt
@@ -1,0 +1,11 @@
+package com.atlassian.performance.tools.jiraactions.api.w3c
+
+/**
+ * Represents the [NavigationType](https://www.w3.org/TR/navigation-timing-2/#dom-navigationtype).
+ */
+enum class NavigationType {
+    NAVIGATE,
+    RELOAD,
+    BACK_FORWARD,
+    PRERENDER;
+}